+++ tests/test_yacman_2.py
@@ -1,44 +1,36 @@
 import os
 
 import pytest
-from jsonschema.exceptions import ValidationError
 
 import yacman
-from yacman.const import FILEPATH_KEY, IK, RO_KEY
+from ubiquerg import read_lock, write_lock
 
 
 class TestWriting:
-    def test_basic_write(self, cfg_file, list_locks, data_path, locked_cfg_file):
-        yacmap = yacman.YAMLConfigManager(filepath=cfg_file)
+    def test_basic_write(self, cfg_file, data_path):
+        yacmap = yacman.YAMLConfigManager.from_yaml_file(cfg_file)
 
-        # Can't write outside a context manager
+        # Can't write outside a write lock
         with pytest.raises(OSError):
             yacmap.write()
 
-        # File is locked when used in a context manager
-        with yacmap as _:
-            assert os.path.exists(locked_cfg_file)
+        # Works inside a write_lock context manager
+        with write_lock(yacmap) as _:
+            yacmap.write()
 
-    def test_write_creates_file(self, data_path, list_locks):
+    def test_write_creates_file(self, data_path):
         yacmap = yacman.YAMLConfigManager(entries={})
         yacmap.write_copy(filepath=make_cfg_file_path("writeout.yaml", data_path))
-
-        # Using .write with a filepath, outside of a context manager,
-        # Doesn't lock the file.
-        # import time
-        # time.sleep(3)
-        # assert os.path.exists(make_lock_path("writeout.yaml", data_path))
-
         assert os.path.exists(make_cfg_file_path("writeout.yaml", data_path))
         os.remove(make_cfg_file_path("writeout.yaml", data_path))
 
     def test_rebase_only_on_locked(self, cfg_file):
-        yacmap = yacman.YAMLConfigManager(filepath=cfg_file)
+        yacmap = yacman.YAMLConfigManager.from_yaml_file(cfg_file)
 
         with pytest.raises(OSError):
             yacmap.rebase()
 
-        with yacmap as y:
+        with read_lock(yacmap) as y:
             y.rebase()
 
 
@@ -48,14 +40,5 @@ class TestReadList:
         print(yacmap.to_yaml())
 
 
-def cleanup_locks(lcks):
-    if lcks:
-        [os.remove(l) for l in lcks]
-
-
 def make_cfg_file_path(name, data_path):
     return os.path.join(data_path, name)
-
-
-def make_lock_path(name, data_path):
-    return os.path.join(data_path, yacman.LOCK_PREFIX + name)